requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
python-dotenv==1.1.0